fix cannot use size (type _Ctype_int) as type _Ctype_ulong (#4212)
* Fixed conversion of unix timestamp to golang time (#3659)

* fix cannot use size (type _Ctype_int) as type _Ctype_ulong

+++ pulsar-client-go/pulsar/c_message.go
@@ -87,8 +87,7 @@ func buildMessage(message ProducerMessage) *C.pulsar_message_t {
 			for i, s := range message.ReplicationClusters {
 				C.setString(array, C.CString(s), C.int(i))
 			}
-
-			C.pulsar_message_set_replication_clusters(cMsg, array, size)
+			C.pulsar_message_set_replication_clusters(cMsg, array, C.size_t(size))
 		}
 	}
